History with chart generated from spider_consolidations. Change some type of migrations.

+++ db/migrate/20121017145000_create_spider_consolidations.rb
@@ -3,8 +3,8 @@ class CreateSpiderConsolidations < ActiveRecord::Migration
     create_table :spider_consolidations do |t|
       t.integer :spider_id
       t.integer :pm_type_axe_id
-      t.integer :average
-      t.integer :average_ref
+      t.float :average
+      t.float :average_ref
       t.integer :ni_number
       t.timestamps
     end
